Bug fixes and quality of life improvements
There seemed to be an issue with the release of redis-py v4.5.2 for socket implementations.  This resulted in errors like `AbstractConnection.__init__() got an unexpected keyword argument socket_timeout` when interacting with redis over a socket.  I have hardcoded versions now to prevent issues like this from randomly popping up.

To allow for faster deployments with changes to config files, Terraform now triggers a replace on the EC2 instance when the config files are changed.